refactor: Moved tooltip rendering logic into a tooltip
(cherry picked from commit c8e5170ddcf4ade4bc6374a6a67db253ec551e77)
